Make join() sequential once sufficient concurrency has been exposed
Spawning and joining tasks has overhead, and thus programs should not
spawn more tasks than required to achieve satisfactory parallel
execution performance. But the optimal number of tasks is system
dependent, and should thus ideally be an easily tunable parameter that the
bulk of a computation is agnostic to.

Here, I experiment with making this tunable parameter fully autotuned.
This works wonderfully in simple cases (10x speedup !), but I expect
I will eventually need to expose the tunable later on for more complex
use cases.